% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RmRMR-auxiliary.R
\name{balancedOrderingDependency}
\alias{balancedOrderingDependency}
\title{Balanced ordering-based dependency measure}
\usage{
balancedOrderingDependency(attrVec, target, numericTarget = NULL,
  nSamples = 100, sampleSize = 10, ...)
}
\arguments{
\item{attrVec, target}{two numeric attributes or an attribute and a decision vector
(not necessarily ordered). At least the first argument has to be numeric.}

\item{numericTarget}{logical indicating whether the \code{target} is ordered and numeric. The 
default is \code{NULL} in which case the function makes a guess using a simple heuristic.}

\item{...}{optional arguments (currently omitted).}
}
\value{
a numeric value expressing ordering dependency between \code{attrVec} and \code{target}
}
\description{
A balanced version of the function for measuring ordering dependency between two vectors 
(the ordering measure).
}
\examples{
#############################################
data(methaneSampleData)

## an experiment on a sample from the data used in a data mining competition - 
## IJCRS'15 Data Challenge: Mining Data from Coal Mines 
## (https://knowledgepit.fedcsis.org/contest/view.php?id=109).
## The whole data set can be downloaded from the competition web page.

mrmrAttrs = mRMRfs(dataT = methaneData$methaneTraining,
                   target = methaneData$methaneTrainingLabels[, V2],
                   dependencyF = balancedOrderingDependency, sampleSize = 10)

mrmrAttrs

}
\author{
Andrzej Janusz
}
\references{
Andrzej Janusz and Marek Grzegorowski. Efficient Attribute Quality Assessment Using 
a Decision Ordering Measure. 2017.
}

