% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RmRMR-auxiliary.R
\name{corrDependency}
\alias{corrDependency}
\title{Computation of indiscernibility classes based on the rough set theory}
\usage{
corrDependency(x, y, ...)
}
\arguments{
\item{x,y}{two numeric attributes or an attribute and a numeric decision
(binary and integer valued decisions are also accepted).}

\item{...}{optional arguments (currently omitted)}
}
\value{
a numeric value expressing linear dependency between \code{x} and \code{y}
}
\description{
This is an exemplary function for measuring inter-attribute dependency
and dependency between attributes and the dacisions. It is used by default
by the provided implementation of the mRMR framework.
}
\examples{
#############################################
data(methaneSampleData)

## an experiment on a sample from methane competition data (IJCRS'16)
mrmrAttrs = mRMRfs(dataT = methaneData$methaneTraining,
                   target = methaneData$methaneTrainingLabels[, as.integer(V2 == 'warning')],
                   dependencyF = corrDependency)

mrmrAttrs
}
\author{
Andrzej Janusz
}
\references{
Mark Hall. Correlation-based Feature Selection for Machine Learning. PhD thesis,
University of Waikato, 1999.
}

