% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RmRMR-data.R
\docType{data}
\name{methaneData}
\alias{methaneData}
\title{Data Sets of the Package}
\format{A list containing 4 tables in data.table format: \code{methaneTraining},
        \code{methaneTest}, \code{methaneTrainingLabels}, \code{methaneTestLabels},
        which are preprocessed subsets of IJCRS’15 Data Challenge training and test data.
        The original time series data were transformed into a feature space using basic statistics.}
\usage{
data(methaneSampleData)
}
\description{
## A sample from the data used in a data mining competition - IJCRS'15 Data Challenge:
## Mining Data from Coal Mines (https://knowledgepit.fedcsis.org/contest/view.php?id=109).
## The whole data set can be downloaded from the competition web page.
}
\references{
Andrzej Janusz, Marek Sikora, Łukasz Wróbel, Sebastian Stawicki,
Marek Grzegorowski, Piotr Wojtas, and Dominik Slezak. Mining Data
from Coal Mines: IJCRS’15 Data Challenge. In Proceedings of RSFDGrC
2015, volume 9437 of LNAI. pp. 429-438, Springer, 2015.
}
\keyword{data}

