% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RmRMR-main.R
\name{mRMRfs}
\alias{mRMRfs}
\title{Feature Subset Selection using mRMR}
\usage{
mRMRfs(dataT, target, dependencyF = corrDependency, Nprobes = 1000,
  allowedRandomness = 0.01, nMax = 20, nCores = 1, ...)
}
\arguments{
\item{dataT}{a data table in \code{data.table} format. Columns of the table shoud
correspond to features (attributes) and rows shoul represent cases (objects).}

\item{target}{a vector of target values for data in \code{dataT}. For the default
dependency function it is assumed that values of \code{target} are numeric,
integer or binary.}

\item{dependencyF}{a function for computing dependencies between attributes and between
attributes and the decisions. The default is an absolute value of
Pearson's correlation (function \code{corrDependency}).}

\item{Nprobes}{an integer specifying the number of probes to use in estimation of
attribute irrelevance (for stopping criteria). The default is \code{1000}.}

\item{allowedRandomness}{a numeric value specifying allowed attribute irrelevance
probability. The default is \code{0.01}.}

\item{nMax}{an integer specifying maximal number of features that can be returned by
the function. The default is \code{20}.}

\item{nCores}{an integer specifying the number of available processor cores for parallel
computations using forking. The default is \code{1} for compatibility with
Windows systems.}

\item{...}{optional arguments (currently omitted).}
}
\value{
an integer vector representing indexes of attributes from the selected subset.
}
\description{
An implementation of the mRMR feature selection freamework for the purpose
of DISESOR project. In this version the random probe test is used as the main
stopping criteria.
}
\examples{
#############################################
data(methaneSampleData)

## an experiment on a sample from methane competition data (IJCRS'16)
mrmrAttrs = mRMRfs(dataT = methaneData$methaneTraining,
                   target = methaneData$methaneTrainingLabels[, as.integer(V2 == 'warning')],
                   dependencyF = corrDependency)

mrmrAttrs

regModel = glm(targets ~.,
               cbind(methaneData$methaneTraining[, mrmrAttrs, with = FALSE],
                     targets = methaneData$methaneTrainingLabels[, as.integer(V2 == 'warning')]),
               family = gaussian(link = "identity"))

preds = predict(regModel, methaneData$methaneTest, type = "response")
caTools::colAUC(preds, methaneData$methaneTestLabels[, V2])
}
\author{
Andrzej Janusz
}
\references{
Hanchuan Peng, Fuhui Long, and Chris Ding. Feature selection based
on mutual information: Criteria of max-dependency, max-relevance, and
min-redundancy. IEEE Trans. Pattern Anal. Mach. Intell., 27(8):1226–1238
}

