% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RmRMR-auxiliary.R
\name{permutationTest}
\alias{permutationTest}
\title{Permutation test of randomness}
\usage{
permutationTest(testedAttr, target, attrScore, dependencyF, Nprobes = 1000,
  nCores = 1, ...)
}
\arguments{
\item{testedAttr}{tested attribute}

\item{target}{the target attribute}

\item{attrScore}{tested attribute's base score value}

\item{dependencyF}{a dependency function used by the test}

\item{Nprobes}{a number of generated random probes}

\item{nCores}{an integer specifying the number of available processor cores for parallel
computations using forking. The default is \code{1} for compatibility with
Windows systems.}

\item{...}{additional parameters (currently omitted)}
}
\value{
a numeric value - an estimation of a probability that a random probe could obtain
        a score equal or greater than \code{attrScore}.
}
\description{
An auxiliary function for performing a permutation test of randomness
}
\examples{
#############################################

}
\author{
Andrzej Janusz
}
\references{
XY
}

