% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RmRMR-auxiliary.R
\name{corrDependency}
\alias{corrDependency}
\title{Linear correlation-based dependency measure}
\usage{
corrDependency(x, y, ...)
}
\arguments{
\item{x, y}{two numeric attributes or an attribute and a numeric decision
(binary and integer valued decisions are also accepted).}

\item{...}{optional arguments to the \code{cor} function.}
}
\value{
a numeric value expressing linear dependency between \code{x} and \code{y}
}
\description{
This is an exemplary function for measuring inter-attribute dependency
and dependency between attributes and the dacisions. It is used by default
by the provided implementation of the mRMR framework.
}
\examples{
#############################################
data(methaneSampleData)

## an experiment on a sample from the data used in a data mining competition - 
## IJCRS'15 Data Challenge: Mining Data from Coal Mines 
## (https://knowledgepit.fedcsis.org/contest/view.php?id=109).
## The whole data set can be downloaded from the competition web page.

mrmrAttrs = mRMRfs(dataT = methaneData$methaneTraining,
                   target = methaneData$methaneTrainingLabels[, as.integer(V2 == 'warning')],
                   dependencyF = corrDependency)

mrmrAttrs

}
\author{
Andrzej Janusz
}
\references{
Mark Hall. Correlation-based Feature Selection for Machine Learning. PhD thesis,
University of Waikato, 1999.
}

