% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RmRMR-data.R
\docType{data}
\name{methaneData}
\alias{methaneData}
\title{Data Sets of the Package}
\format{A list containing 4 tables in data.table format: \code{methaneTraining},
        \code{methaneTest}, \code{methaneTrainingLabels}, \code{methaneTestLabels},
        which are subsets of IJCRS’15 Data Challenge training and test data.}
\usage{
data(methaneSampleData)
}
\description{
Data Sets of the Package
}
\references{
Andrzej Janusz, Marek Sikora, Łukasz Wróbel, Sebastian Stawicki,
Marek Grzegorowski, Piotr Wojtas, and Dominik Slezak. Mining Data
from Coal Mines: IJCRS’15 Data Challenge. In Proceedings of RSFDGrC
2015, volume 9437 of LNAI. Springer, 2015.
}
\keyword{data}

